Deduplicate common database_logic
